monitor/pyproject.toml
[tool.poetry]
name = "monitor"
version = "0.1.0"
description = ""
authors = ["ricardodiaz <ridiazx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
